LayoutTests/media/modern-media-controls/mute-button/mute-button-expected.txt

Testing the MuteButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS muteButton.element.localName is "button"
PASS muteButton.element.classList.contains("icon") is true
PASS muteButton.element.classList.contains("mute") is true
PASS muteButton.iconName is "volume"
PASS muteButton.muted is false
PASS mutedButton.muted is true
PASS mutedButton.iconName is "volume-mute"
PASS muteButton.element.style.webkitMaskImage.includes("macOS/volume@") became true
PASS mutedButton.element.style.webkitMaskImage.includes("macOS/volume-mute@") became true
PASS successfullyParsed is true

TEST COMPLETE
